telegram/requirements.txt
aiogram
fastapi
uvicorn
python-dotenv
openai
